Re-enable QGraphicsView test on Windows, Mac and Ubuntu 11.10.
On Windows, the test had a stable failure and an XPASS related to an
already closed bug report.  On Mac, the test had one stable failure, on
the same statement that had the XPASS on Windows.  On Ubuntu, the test
has been passing consistently.

Mark the stable failures with QEXPECT_FAIL, remove the QEXPECT_FAIL that
causes the XPASS, and remove the insignificant_test markers to re-enable
the test.

Task-number: QTBUG-24296
Change-Id: I05dd16b39296919268fee216332929df591f2b66
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>